refactor(renderer/test): split delimited block tests (part 2)
Splitting more files, using a common
`delimited_block_` prefix on all files so they are
easier to locate in the project explorer

Follow-up of #776
Updates #775
