+++ pkg/renderer/sgml/delimited_block_listing.go
@@ -1,0 +1,48 @@
+package sgml
+
+import (
+	"strings"
+
+	"github.com/bytesparadise/libasciidoc/pkg/renderer"
+	"github.com/bytesparadise/libasciidoc/pkg/types"
+	"github.com/pkg/errors"
+)
+
+func (r *sgmlRenderer) renderListingBlock(ctx *renderer.Context, b types.ListingBlock) (string, error) {
+	if k, found := b.Attributes[types.AttrKind]; found && k == types.Source {
+		return r.renderSourceBlock(ctx, b)
+	}
+	previousWithinDelimitedBlock := ctx.WithinDelimitedBlock
+	previousIncludeBlankLine := ctx.IncludeBlankLine
+	defer func() {
+		ctx.WithinDelimitedBlock = previousWithinDelimitedBlock
+		ctx.IncludeBlankLine = previousIncludeBlankLine
+	}()
+	ctx.WithinDelimitedBlock = true
+	ctx.IncludeBlankLine = true
+	result := &strings.Builder{}
+	lines := discardEmptyLines(b.Lines)
+	content, err := r.renderLines(ctx, lines)
+	if err != nil {
+		return "", errors.Wrap(err, "unable to render listing block content")
+	}
+	roles, err := r.renderElementRoles(ctx, b.Attributes)
+	if err != nil {
+		return "", errors.Wrap(err, "unable to render listing block roles")
+	}
+
+	err = r.listingBlock.Execute(result, struct {
+		Context *renderer.Context
+		ID      string
+		Title   string
+		Roles   string
+		Content string
+	}{
+		Context: ctx,
+		ID:      r.renderElementID(b.Attributes),
+		Title:   r.renderElementTitle(b.Attributes),
+		Roles:   roles,
+		Content: content,
+	})
+	return result.String(), err
+}
