+++ pkg/renderer/sgml/delimited_block_verse.go
@@ -1,0 +1,43 @@
+package sgml
+
+import (
+	"strings"
+
+	"github.com/bytesparadise/libasciidoc/pkg/renderer"
+	"github.com/bytesparadise/libasciidoc/pkg/types"
+	"github.com/pkg/errors"
+)
+
+func (r *sgmlRenderer) renderVerseBlock(ctx *renderer.Context, b types.VerseBlock) (string, error) {
+	result := &strings.Builder{}
+	roles, err := r.renderElementRoles(ctx, b.Attributes)
+	if err != nil {
+		return "", errors.Wrap(err, "unable to render verser block")
+	}
+	previousWithinDelimitedBlock := ctx.WithinDelimitedBlock
+	defer func() {
+		ctx.WithinDelimitedBlock = previousWithinDelimitedBlock
+	}()
+	ctx.WithinDelimitedBlock = true
+	content, err := r.renderLines(ctx, discardEmptyLines(b.Lines))
+	if err != nil {
+		return "", errors.Wrap(err, "unable to render verse block")
+	}
+
+	err = r.verseBlock.Execute(result, struct {
+		Context     *renderer.Context
+		ID          string
+		Title       string
+		Roles       string
+		Attribution Attribution
+		Content     string
+	}{
+		Context:     ctx,
+		ID:          r.renderElementID(b.Attributes),
+		Title:       r.renderElementTitle(b.Attributes),
+		Roles:       roles,
+		Attribution: verseBlockAttribution(b),
+		Content:     string(content),
+	})
+	return result.String(), err
+}
