+++ pkg/renderer/sgml/html5/delimited_block_listing_tmpl.go
@@ -1,0 +1,11 @@
+package html5
+
+const (
+	listingBlockTmpl = `<div {{ if .ID }}id="{{ .ID }}" {{ end }}` +
+		"class=\"listingblock{{ if .Roles }} {{ .Roles }}{{ end }}\">\n" +
+		"{{ if .Title }}<div class=\"title\">{{ .Title }}</div>\n{{ end }}" +
+		"<div class=\"content\">\n" +
+		"<pre>{{ .Content }}</pre>\n" +
+		"</div>\n" +
+		"</div>\n"
+)
