+++ pkg/renderer/sgml/html5/delimited_block_source_tmpl.go
@@ -1,0 +1,16 @@
+package html5
+
+const (
+	sourceBlockTmpl = `<div {{ if .ID }}id="{{ .ID }}" {{ end }}` +
+		"class=\"listingblock{{ if .Roles }} {{ .Roles }}{{ end }}\">\n" +
+		"{{ if .Title }}<div class=\"title\">{{ .Title }}</div>\n{{ end }}" +
+		"<div class=\"content\">\n" +
+		"<pre class=\"" +
+		`{{ if .SyntaxHighlighter }}{{ .SyntaxHighlighter }} {{ end }}` +
+		`highlight">` +
+		`<code{{ if .Language }}{{ if not .SyntaxHighlighter }} class="language-{{ .Language}}"{{ end }} ` +
+		`data-lang="{{ .Language}}"{{ end }}>` +
+		"{{ .Content }}</code></pre>\n" +
+		"</div>\n" +
+		"</div>\n"
+)
