+++ pkg/renderer/sgml/html5/delimited_block_verse_tmpl.go
@@ -1,0 +1,12 @@
+package html5
+
+const (
+	verseBlockTmpl = `<div {{ if .ID }}id="{{ .ID }}" {{ end }}` +
+		"class=\"verseblock{{ if .Roles }} {{ .Roles }}{{ end }}\">\n" +
+		"{{ if .Title }}<div class=\"title\">{{ .Title }}</div>\n{{ end }}" +
+		"<pre class=\"content\">{{ .Content }}</pre>\n" +
+		"{{ if .Attribution.First }}<div class=\"attribution\">\n&#8212; {{ .Attribution.First }}" +
+		"{{ if .Attribution.Second }}<br>\n<cite>{{ .Attribution.Second }}</cite>{{ end }}\n" +
+		"</div>\n{{ end }}" +
+		"</div>\n"
+)
